requirements.txt
diff-match-patch==20121119
Django==1.11.4
django-graphos==0.3.41
django-import-export==0.5.1
et-xmlfile==1.0.1
jdcal==1.3
jsonfield==2.0.2
mysqlclient==1.3.12
odfpy==1.3.5
openpyxl==2.4.8
pytz==2017.2
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
